% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{body_get_request}
\alias{body_get_request}
\title{Builds the XML body to insert data}
\usage{
body_get_request(.user, .type, .branch, .instance)
}
\arguments{
\item{.user}{username}

\item{.type}{data source, EBXCodelist or EBXGroup}

\item{.branch}{branch name}

\item{.instance}{intance name}
}
\value{
character
}
\description{
Builds the XML body to insert data
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
