% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{body_remove_request}
\alias{body_remove_request}
\title{Builds the XML body to remove data}
\usage{
body_remove_request(.user, .cl_name, .folder = "Metadata", .branch,
  .instance)
}
\arguments{
\item{.user}{username}

\item{.cl_name}{code list name which the data will be read from. Please, see
the code list options by running the function \code{\link{GetEBXCodeLists}}}

\item{.folder}{folder name}

\item{.branch}{branch name}

\item{.instance}{intance name}
}
\value{
character
}
\description{
Builds the XML body to remove data
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
