% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{body_request_data}
\alias{body_request_data}
\title{Builds the XML body with data}
\usage{
body_request_data(.user, .secret, .branch, .instance, .folder, .folder2,
  .table, .verb)
}
\arguments{
\item{.user}{username}

\item{.secret}{user password}

\item{.branch}{branch name}

\item{.instance}{intance name}

\item{.folder}{folder name}

\item{.folder2}{optinal second level folder}

\item{.table}{the table name}

\item{.verb}{SOAP action verb}
}
\value{
character
}
\description{
Builds the XML body which containd data.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
