% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{body_select_request}
\alias{body_select_request}
\title{Builds the XML body to select data}
\usage{
body_select_request(.user, .name, .branch, .instance)
}
\arguments{
\item{.user}{username}

\item{.name}{code list or group name.}

\item{.branch}{branch name}

\item{.instance}{intance name}
}
\value{
character
}
\description{
Builds the XML body to select data
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
