% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{body_update_request}
\alias{body_update_request}
\title{Builds the XML body to update data}
\usage{
body_update_request(.user, .name, .folder = "Metadata", .branch,
  .instance)
}
\arguments{
\item{.user}{username}

\item{.name}{code list or group name.}

\item{.folder}{folder name}

\item{.branch}{branch name}

\item{.instance}{intance name}
}
\value{
character
}
\description{
Builds the XML body to update data
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
