% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cl_data_insert_xml}
\alias{cl_data_insert_xml}
\title{Insert XML; internal}
\usage{
cl_data_insert_xml(.data, .table = "EBXCodelist")
}
\arguments{
\item{.data}{data for request}

\item{.table}{table name in EBX}
}
\value{
XML character
}
\description{
Create XML for soap insert request
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
