% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_convert_xml}
\alias{data_convert_xml}
\title{Convert to XML; internal}
\usage{
data_convert_xml(.data, .table = .table)
}
\arguments{
\item{.data}{data for request}

\item{.table}{table name in EBX}
}
\value{
XML character
}
\description{
Create XML for soap request
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
