% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBX_Insert.R
\name{EBXInsert}
\alias{EBXInsert}
\title{Insert rows in EBX5 Code List}
\usage{
EBXInsert(branch, instance, folder, folder2 = "", table, data,
  connection = NA)
}
\arguments{
\item{branch}{dataset branch}

\item{instance}{dataset instance}

\item{folder}{folder where the table is}

\item{folder2}{optional second level folder}

\item{table}{ebx5 table name}

\item{data}{a \code{\link[base]{data.frame}} containing all columns to be be inserted}

\item{connection}{the result of \code{\link{GetEBXConnection}}; optional}
}
\value{
boolean
}
\description{
This function will insert data rows into a code list stored in EBX5.
}
\details{
The data columns provided must follow the code-list which is being inserted.
All fiels of a primary key (usually the Identifier), and all mandatory fields must be present.
Using the SOAP-API the column names and tabe table name are not the label visible in EBX, but the name.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\examples{

\dontrun{
cl_insert <- data.frame(Identifier = c(999, 888), Name_En=c('Test1','Test2'))
EBXInsert(branch='Fishery',
          instance='Fishery',
          folder='Metadata',
          table='Test_table',
          data=cl_insert)
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
