% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBX_Read.R
\name{EBXRead}
\alias{EBXRead}
\title{Read EBX5 Code List}
\usage{
EBXRead(branch, instance, folder, folder2 = "", table, connection = NA)
}
\arguments{
\item{branch}{dataset branch}

\item{instance}{dataset instance}

\item{folder}{folder where the table is}

\item{folder2}{optional second level folder}

\item{table}{ebx5 table name}

\item{connection}{the result of \code{\link{GetEBXConnection}}; optional}
}
\value{
Return an object of the class \code{\link[data.table]{data.table}}.
}
\description{
This function reads a table from EBX5.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\details{
Using the SOAP-API the tabe table name is not the label visible in EBX, but the name.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\examples{
\dontrun{
EBXRead(branch='Fishery',
          instance='Fishery',
          folder='Metadata',
          table='Test_table')
}
}
\seealso{
\code{\link{GetEBXConnection}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
