% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBX_Update.R
\name{EBXUpdate}
\alias{EBXUpdate}
\title{Update rows in EBX5 Code List}
\usage{
EBXUpdate(branch, instance, folder, folder2 = "", table, data, connection = NA)
}
\arguments{
\item{branch}{dataset branch}

\item{instance}{dataset instance}

\item{folder}{folder where the table is}

\item{folder2}{optional second level folder}

\item{table}{ebx5 table name}

\item{data}{a \code{\link[base]{data.frame}} containing the columns to be be updated}

\item{connection}{the result of \code{\link{GetEBXConnection}}; optional}
}
\value{
boolean
}
\description{
This function updates existing data rows of a code list stored in EBX5.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\details{
The data columns provided must follow the code-list which is being updated.
All fiels of a primary key (usually the Identifier) must be present.
Using the SOAP-API the column names and tabe table name are not the label visible in EBX, but the name.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\examples{
\dontrun{
cl_update <- data.frame(
               Identifier = c(999, 888),
               Acronym = 'TEST_ACRONYM',
               Folder = 'TESTFOLDER',
               Name = 'TEST_NAME',
               Branch = 'Fishery',
               Instance = 'Fishery')
EBXUpdate(branch   = 'Fishery',
          instance = 'Fishery',
          folder   = 'Metadata',
          table    = 'Test_table',
          data     = cl_update)
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
