% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoebx5.R
\docType{package}
\name{faoebx5}
\alias{faoebx5}
\alias{faoebx5-package}
\title{faoebx5: an R package to access EBX5 data from R.}
\description{
The faoebx5 package provides tools to read, write,
and update data in the EBX5 through R.
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}

Thomas Berger, \email{thomas.berger@fao.org}
}
