% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEBX_Table}
\alias{getEBX_Table}
\title{get table; internal}
\usage{
getEBX_Table(connection, branch, instance, table_name)
}
\arguments{
\item{connection}{connection details}

\item{branch}{table branch}

\item{instance}{table instance}

\item{table_name}{table name}
}
\value{
character XML
}
\description{
Reads a table from EBX
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
