% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_GetEBXCodeLists.R
\name{GetEBXCodeLists}
\alias{GetEBXCodeLists}
\title{Get EBX5 Code List}
\usage{
GetEBXCodeLists(branch = "Fishery", instance = "Fishery")
}
\arguments{
\item{branch}{branch name.}

\item{instance}{instance name.}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function aimed to get the list of code lists
available in the EBX5 to be read from using the function
\code{\link{ReadEBXCodeList}}.
}
\examples{

\dontrun{
GetEBXCodeLists()
}

}
\seealso{
\code{\link{ReadEBXCodeList}}
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
