% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_GetEBXGroups.R
\name{GetEBXGroups}
\alias{GetEBXGroups}
\title{Get EBX Groups}
\usage{
GetEBXGroups(branch = "Fishery", instance = "Fishery")
}
\arguments{
\item{branch}{branch name.}

\item{instance}{instance name.}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function aimed to get the list of groups
available in the EBX5 to be read from using the function
\code{\link{ReadEBXGroup}}.
}
\examples{

\dontrun{
GetEBXGroups()
}
}
\seealso{
\code{\link{ReadEBXGroup}}
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
