% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_GetSecret.R
\name{GetSecret}
\alias{GetSecret}
\title{Get EBX Secret}
\usage{
GetSecret(username)
}
\arguments{
\item{username}{EBX username, character}
}
\value{
the username secret stored
}
\description{
Get EBX Secret
}
