% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{header_fields}
\alias{header_fields}
\title{Header Fields}
\usage{
header_fields(soap_action)
}
\arguments{
\item{soap_action}{EBX5 SOAP URL to API requests}
}
\value{
character with the header fields
}
\description{
Builds the header fields
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
