% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_ReadEBXCodeList.R
\name{ReadEBXCodeList}
\alias{ReadEBXCodeList}
\title{Read EBX5 Code List}
\usage{
ReadEBXCodeList(cl_name, branch = "Fishery", instance = "Fishery")
}
\arguments{
\item{cl_name}{code list name which the data will be read from. Please, see
the code list options by running the function \code{\link{GetEBXCodeLists}} in the field "Name".}

\item{branch}{branch name.}

\item{instance}{instance name.}
}
\value{
Return an object of the class \code{\link[data.table]{data.table}}.
}
\description{
This function aimed to read code list data from EBX5 to R.
}
\examples{
\dontrun{
ReadEBXCodeList(cl_name = "ISSCFC")
}
}
\seealso{
\code{\link{GetEBXCodeLists}}.
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
