% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_ReadEBXGroup.R
\name{ReadEBXGroup}
\alias{ReadEBXGroup}
\title{Read EBX5 Group}
\usage{
ReadEBXGroup(gr_name, branch = "Fishery", instance = "Fishery")
}
\arguments{
\item{gr_name}{group name which the data will be read from.  Please, see
the code list options after run the function \code{\link{GetEBXGroups}} in the field "Name".}

\item{branch}{branch name.}

\item{instance}{instance name.}
}
\value{
Return an object of the class \code{\link[data.table]{data.table}}.
}
\description{
This function aimed to read group data from EBX5 to R.
}
\examples{

\dontrun{
gr <- ReadEBXGroup(gr_name = 'Group_CPCDiv_CPCGroup')
}

}
\seealso{
\code{\link{GetEBXGroups}}.
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
