% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_RemoveEBXCodeList.R
\name{RemoveEBXCodeList}
\alias{RemoveEBXCodeList}
\title{Remove rows in EBX5 Code List}
\usage{
RemoveEBXCodeList(data, cl_name, folder, branch = "Fishery",
  instance = "Fishery")
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{cl_name}{code list name which the data will be read from. Please, see
the code list options by running the function \code{\link{GetEBXCodeLists}} in the field "Name".}

\item{folder}{folder name in EBX that the code list is stored.  Please, see
the code list options by running the function \code{\link{GetEBXCodeLists}} (\code{\link[base]{character}}).}

\item{branch}{branch name.}

\item{instance}{instance name.}
}
\value{
boolean
}
\description{
This function aimed to remove
data rows into a code list stored in EBX5 through R.
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
