% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_RemoveEBXCredentials.R
\name{RemoveEBXCredentials}
\alias{RemoveEBXCredentials}
\title{Remove EBX Credentials}
\usage{
RemoveEBXCredentials(username)
}
\arguments{
\item{username}{username, a character scalar.}
}
\value{
Status 0 (zero) ok.
}
\description{
This function aimed to remove your EBX credentials previously
created using the function \code{\link{SetEBXCredentials}}.
}
\details{
When you run the function \code{RemoveEBXCredentials()} will open
a box to type your password twice. This function tries to guess your username
using the variable stored in the system, Sys.getenv('USERNAME'). Otherwise,
you can provide your username by the argument \code{username}.
}
\seealso{
\code{\link{SetEBXCredentials}}.
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
credential stored in the EBX file
credentials stored in the default variable
}
