% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_RemoveEBXGroup.R
\name{RemoveEBXGroup}
\alias{RemoveEBXGroup}
\title{Remove rows in EBX5 Group}
\usage{
RemoveEBXGroup(data, gr_name, folder, branch = "Fishery",
  instance = "Fishery")
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{gr_name}{group name which the data will be read from.  Please, see
the code list options after run the function \code{\link{GetEBXGroups}} in the field "Name".}

\item{folder}{folder name in EBX that the code list is stored.  Please, see
the code list options by running the function \code{\link{GetEBXGroups}}}

\item{branch}{branch name.}

\item{instance}{instance name.}
}
\value{
boolean
}
\description{
This function aimed to remove
data rows into a group stored in EBX5 through R.
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
