% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_SetEBXCredentials.R
\name{SetEBXCredentials}
\alias{SetEBXCredentials}
\title{Set EBX Credentials}
\usage{
SetEBXCredentials(username, password, lock = FALSE, new = TRUE)
}
\arguments{
\item{username}{username, a character.}

\item{password}{password, a character.}

\item{lock}{logical, default is FALSE. If it is TRUE, will store a locked file with the password,
and whenever trying to use the credential will ask for the password to unlock the data.}

\item{new}{logical, default is TRUE. It indicates whether will be set credentials to a new user.}
}
\value{
Status 0 (zero) ok.
}
\description{
This function aimed to set up your EBX credentials.
First of all, you have to request the EBX manager to give rights
to access EBX. Usually, you will use the same username and password
from the Statistical Working System (SWS).
}
\details{
When you run the function \code{SetEBXCredentials()} will open
a box to type your password twice. This function tries to guess your username
using the variable stored in the system, Sys.getenv('USERNAME'). Otherwise,
you can provide your username by the argument \code{username}.
}
\seealso{
\code{\link{RemoveEBXCredentials}}.
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
