% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soap_request}
\alias{soap_request}
\title{Builds the XML body to insert data}
\usage{
soap_request(.user, .secret, .verb, .type, .branch, .instance)
}
\arguments{
\item{.type}{data source, EBXCodelist or EBXGroup}
}
\value{
character
}
\description{
Builds the XML body to insert data
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
