% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soap_request_read}
\alias{soap_request_read}
\title{Builds the XML body to insert data; internal}
\usage{
soap_request_read(.user, .secret, .verb, .table, .branch, .instance)
}
\arguments{
\item{.user}{EBX5 user name}

\item{.secret}{name secret}

\item{.verb}{action (select, update, insert, delete)}

\item{.table}{the code-list name}

\item{.branch}{dataset branch}

\item{.instance}{dataset instance}
}
\value{
character XML
}
\description{
Builds the XML body to insert data
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
