% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update}
\alias{update}
\title{update; internal}
\usage{
update(data, sdmx_name, verb, isCodeList = TRUE)
}
\arguments{
\item{data}{user data}

\item{sdmx_name}{SDMX table name}

\item{verb}{what todo (select, insert, update)}

\item{isCodeList}{TRUE uses ebx5.cl_datal FALSE uses ebx5.gr_data}

\item{instance}{table instance}
}
\value{
TRUE if update was done
}
\description{
Updates a table
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
