% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_UpdateEBXGroup.R
\name{UpdateEBXGroup}
\alias{UpdateEBXGroup}
\title{Update rows in EBX5 Group}
\usage{
UpdateEBXGroup(data, gr_name, folder, branch = "Fishery",
  instance = "Fishery")
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{gr_name}{group name which the data will be read from.  Please, see
the code list options after run the function \code{\link{GetEBXGroups}} in the field "Name".}

\item{folder}{folder name in EBX that the code list is stored.  Please, see
the code list options by running the function \code{\link{GetEBXGroups}}}

\item{branch}{branch name.}

\item{instance}{instance name.}
}
\value{
boolean
}
\description{
This function aimed to update
data rows into a group stored in EBX5 through R.
}
\details{
Note that the udpated rows must have the same columns name os the table that will be updated.
}
\examples{

\dontrun{
UpdateEBXGroup(data = gr_update,
gr_name = 'EBXGroup',
folder = 'Metadata')
}

}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
