% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateTable}
\alias{updateTable}
\title{update; internal}
\usage{
updateTable(connection, branch, instance, folder, data, table, verb)
}
\arguments{
\item{connection}{connection details}

\item{branch}{table branch}

\item{instance}{table instance}

\item{verb}{what todo (select, insert, update)}

\item{table_name}{table name}
}
\value{
TRUE if update was done
}
\description{
Updates a table
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
