% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cl_data_insert_xml}
\alias{cl_data_insert_xml}
\title{XXX}
\usage{
cl_data_insert_xml(.data, .cl_name = "EBXCodelist")
}
\arguments{
\item{.data}{XXX}

\item{.cl_name}{code list name which the data will be read from. Please, see
the code list options by running the function \code{\link{GetEBXCodeLists}}}
}
\value{
XXX
}
\description{
XXX
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
