% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBX_Remove.R
\name{EBXRemove}
\alias{EBXRemove}
\title{Remove rows in EBX5 Code List}
\usage{
EBXRemove(branch, instance, table, data, connection = NA)
}
\arguments{
\item{branch}{dataset branch}

\item{instance}{dataset instance}

\item{table}{ebx5 table name}

\item{data}{a \code{\link[base]{data.frame}} that will be removed.
data must specify the unique primary key(s); which identify the row to be removed.}

\item{connection}{the result of \code{\link{GetEBXConnection}}; optional}
}
\value{
boolean
}
\description{
This function deletes a single data row from a code list stored in EBX5.
}
\details{
All fields of a primary key (usually the Identifier) must be present.
Using the SOAP-API the column names and tabe table name are not the label visible in EBX, but the name.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\examples{
\dontrun{
cl_remove <- data.frame(
Identifier = c(999))

EBXRemove( branch   = 'Fishery',
          instance = 'Fishery',
          table    = 'Test_table',
          data     = cl_remove)
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
