% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_GetEBXCodeLists.R
\name{GetEBXCodeLists}
\alias{GetEBXCodeLists}
\title{Get EBX5 Code Lists}
\usage{
GetEBXCodeLists(connection)
}
\arguments{
\item{connection}{is optional}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the list of code lists
defined in the 'Metatadata' structure in EBX5. Metadata contains
the instance, branch and the code-list name used in EBX5.
This way, the software uses the SDMX-style codelist name, and the actual
location inside EBX5 can be changed without breaking the software.
\code{\link{ReadEBXCodeList}}.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\examples{

\dontrun{
GetEBXCodeLists()
}

}
\seealso{
\code{\link{ReadEBXCodeList}} \code{\link{GetEBXGroups}}
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
