% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_GetEBXConnection.R
\name{GetEBXConnection}
\alias{GetEBXConnection}
\title{Get EBX5 Connection details}
\usage{
GetEBXConnection()
}
\value{
Returns a \code{\link[data.table]{data.table}} with the connection details
}
\description{
This is a utility function to return connection details
required for using the EBX5 SOAP web-API. This function is not intended for
users, it is exported beacuse it is useful for test purposes.
}
\seealso{
\code{\link{SetupEBXConnection}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
