% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_GetEBXGroups.R
\name{GetEBXGroups}
\alias{GetEBXGroups}
\title{Get EBX Groups}
\usage{
GetEBXGroups(connection)
}
\arguments{
\item{connection}{is optional}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the list of groups
defined in the 'Metatadata' structure in EBX5. Metadata contains
the instance, branch and the code-list name used in EBX5.
This way, the software uses the SDMX-style group name, and the actual
location inside EBX5 can be changed without breaking the software.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\examples{

\dontrun{
GetEBXGroups()
}
}
\seealso{
\code{\link{ReadEBXGroup}} \code{\link{GetEBXCodeLists}}
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
