% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{header_fields}
\alias{header_fields}
\title{Soap header; internal}
\usage{
header_fields(soap_verb)
}
\arguments{
\item{soap_verb}{EBX5 SOAP verb}
}
\value{
XML character with the header fields
}
\description{
Create the XML header
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
