% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_InsertEBXGroup.R
\name{InsertEBXGroup}
\alias{InsertEBXGroup}
\title{Insert rows in EBX5 Group}
\usage{
InsertEBXGroup(data, sdmx_group_name)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{sdmx_group_name}{group name, in SDMX style.
Available code lists are shown by function \code{\link{GetEBXGroups}} in the field "Acronym".
The actual group location in EBX5 (branch, instance, code-list-name) are resolved
using the metadata structure.}
}
\value{
boolean
}
\description{
This function inserts data rows into a group stored in EBX5.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\details{
Note that the new rows must have the same columns name os the table that will be appended.
}
\examples{

\dontrun{
InsertEBXGroup(data = gr_new,
gr_name  = 'EBXGroup',
folder   = 'Metadata',
branch   = 'Fishery',
instance = 'Fishery')
}

}
\seealso{
\code{\link{GetEBXGroups}}
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
