% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_InsertEBXGroup.R
\name{InsertEBXGroup}
\alias{InsertEBXGroup}
\title{Insert rows in EBX5 Group}
\usage{
InsertEBXGroup(data, gr_name, folder, branch = "Fishery",
  instance = "Fishery")
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{folder}{folder name in EBX that the code list is stored.  Please, see
the code list options by running the function \code{\link{GetEBXGroups}}}
}
\value{
boolean
}
\description{
This function aimed to insert
data rows into a group stored in EBX5 through R.
}
\details{
Note that the new rows must have the same columns name os the table that will be appended.
}
\examples{

\dontrun{
InsertEBXGroup(data = gr_new,
gr_name  = 'EBXGroup',
folder   = 'Metadata',
branch   = 'Fishery',
instance = 'Fishery')
}

}
\seealso{
\code{\link{GetEBXGroups}}
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
