% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_ReadEBXCodeList.R
\name{ReadEBXCodeList}
\alias{ReadEBXCodeList}
\title{Read EBX5 Code List}
\usage{
ReadEBXCodeList(sdmx_name)
}
\arguments{
\item{sdmx_name}{code list name which the data will be read from. Please, see
the code list options by running the function \code{\link{GetEBXCodeLists}} in the field "Acronym".}
}
\value{
Return an object of the class \code{\link[data.table]{data.table}}.
}
\description{
This function aimed to read code list data from EBX5 to R.
}
\examples{
\dontrun{
ReadEBXCodeList(sdmx_name = "CL_FI_COMMODITY_ISSCFC")
}
}
\seealso{
\code{\link{GetEBXCodeLists}}.
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
