% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_ReadEBXCodeList.R
\name{ReadEBXCodeList}
\alias{ReadEBXCodeList}
\title{Read EBX5 Code List}
\usage{
ReadEBXCodeList(sdmx_code_list_name)
}
\arguments{
\item{sdmx_code_list_name}{code list name, in SDMX style. Please, see
the code lists available by running the function \code{\link{GetEBXCodeLists}} in the field "Acronym".
The actual codelist location in EBX5 (branch, instance, code-list-name) are resolved
using the metadata structure.}
}
\value{
Return an object of the class \code{\link[data.table]{data.table}}.
}
\description{
This function reads a code list from EBX5 to R.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\examples{
\dontrun{
ReadEBXCodeList(sdmx_name = "CL_FI_COMMODITY_ISSCFC")
}
}
\seealso{
\code{\link{GetEBXCodeLists}}.
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
