% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_ReadEBXGroup.R
\name{ReadEBXGroup}
\alias{ReadEBXGroup}
\title{Read EBX5 Group}
\usage{
ReadEBXGroup(sdmx_name)
}
\arguments{
\item{sdmx_name}{group name, in SDMX style. Please, see
the groups available by running the function \code{\link{GetEBXGroups}} in the field "Acronym".
The actual group location in EBX5 (branch, instance, code-list-name) are resolved
using the metadata structure.}
}
\value{
Return an object of the class \code{\link[data.table]{data.table}}.
}
\description{
This function reads group data from EBX5 to R.
}
\examples{

\dontrun{
gr <- ReadEBXGroup(sdmx_name = 'HCL_FI_COMMODITY_FAOL1_FAOL2')
}

}
\seealso{
\code{\link{GetEBXGroups}}.
}
\author{
Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
