% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_RemoveEBXCodeList.R
\name{RemoveEBXCodeList}
\alias{RemoveEBXCodeList}
\title{Remove rows in EBX5 Code List}
\usage{
RemoveEBXCodeList(data, sdmx_codelist_name)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be inserted.}

\item{sdmx_codelist_name}{codelist name see Acronym \code{\link{GetEBXCodeLists}} (\code{\link[base]{character}}).}
}
\value{
boolean
}
\description{
This function deletes data rows from a code list stored in EBX5.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\details{
For a delete operation, data must specify all keys of the code-list to uniqely idetify the rows.
Using the SOAP-API the column name is not the label visible in EBX, but the field name.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
