% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_RemoveEBXCodeList.R
\name{RemoveEBXCodeList}
\alias{RemoveEBXCodeList}
\title{Remove rows in EBX5 Code List}
\usage{
RemoveEBXCodeList(data, sdmx_name)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{sdmx_name}{code list name, in SDMX style. Please, see
the code lists available by running the function \code{\link{GetEBXCodeLists}} in the field "Acronym".
The actual codelist location in EBX5 (branch, instance, code-list-name) are resolved
using the metadata structure.}
}
\value{
boolean
}
\description{
This function aimed to remove
data rows into a code list stored in EBX5 through R.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
