% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_RemoveEBXGroup.R
\name{RemoveEBXGroup}
\alias{RemoveEBXGroup}
\title{Remove rows in EBX5 Group}
\usage{
RemoveEBXGroup(data, sdmx_name)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}
}
\value{
boolean
}
\description{
This function removes data rows from a group stored in EBX5.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
