% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_SetEBXCredentials.R
\name{SetEBXCredentials}
\alias{SetEBXCredentials}
\title{Set EBX Credentials}
\usage{
SetEBXCredentials(username)
}
\arguments{
\item{username}{username, a character scalar.}
}
\value{
Status 0 (zero) ok.
}
\description{
This function aimed to set up your EBX credentials.
First of all, you have to request the EBX manager to give rights
to access EBX. Usually, you will use the same username and password
from the Statistical Working System (SWS).
}
\details{
When you run the function \code{SetEBXCredentials()} will open
a box to type your password twice. This function tries to guess your username
using the variable stored in the system, Sys.getenv('USERNAME'). Otherwise,
you can provide your username by the argument \code{username}.
}
\seealso{
\code{\link{RemoveEBXCredentials}}.
}
\author{
Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
