% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_SetupEBX Connection.R
\name{SetupEBXConnection}
\alias{SetupEBXConnection}
\title{Setup EBX Connection}
\usage{
SetupEBXConnection(meta_branch, meta_instance, ebx_soap_url, username,
  password)
}
\arguments{
\item{meta_branch}{metadata branch name in EBX}

\item{meta_instance}{metadata instance name in EBX}

\item{ebx_soap_url}{SOAP webservice url}

\item{username}{username, a character.}

\item{password}{password, a character.}

\item{lock}{logical, default is FALSE. If it is TRUE, will store a locked file with the password,
and whenever trying to use the credential will ask for the password to unlock the connection setup.}
}
\value{
Status 0 (zero) ok.
}
\description{
This function will store the SOAP connection details in a yaml file
}
\seealso{
\code{\link{RemoveEBXConnection}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
