% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_SetupEBX_Connection.R
\name{SetupEBXConnection}
\alias{SetupEBXConnection}
\title{Setup EBX Connection}
\usage{
SetupEBXConnection(
  meta_branch = "Fishery",
  meta_instance = "Fishery",
  ebx_soap_url,
  username,
  password
)
}
\arguments{
\item{meta_branch}{metadata branch name in EBX, used for faostatR library (default is Fishery)}

\item{meta_instance}{metadata instance name in EBX, used for faostatR library (default is Fishery)}

\item{ebx_soap_url}{SOAP webservice url}

\item{username}{username, a character.}

\item{password}{password, a character.}
}
\value{
Status 0 (zero) ok.
}
\description{
This function will store the SOAP connection details in a yaml file
}
\details{
In case of an shiny app, run the \code{\link{SetupEBXConnection}} from your Rstudio.
The conncetion details are persisted to a file [ebx5_connecion.yml] which you can then embed into the shiny app.
The library faoebx5 uses \code{\link{GetEBXConnection}}, in order to obtain the connection details from the yaml file.
Obviously, the yaml file is excluded, not part of the GIT repository faoebx5.
}
\seealso{
\code{\link{GetEBXConnection}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
