% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_UpdateEBXCodeList.R
\name{UpdateEBXCodeList}
\alias{UpdateEBXCodeList}
\title{Update rows in EBX5 Code List}
\usage{
UpdateEBXCodeList(data, sdmx_codelist_name)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} containing all columns to be be updated}

\item{sdmx_codelist_name}{code list name, in SDMX style.
Available code lists are shown by function \code{\link{GetEBXCodeLists}} in the field "Acronym".
The actual codelist location in EBX5 (branch, instance, code-list-name) are resolved
using the metadata structure.}
}
\value{
boolean
}
\description{
This function updates data rows of a code list stored in EBX5.
Requires that the EBX5 connection was configured using \code{\link{SetupEBXConnection}}.
}
\details{
Note that the udpated rows must have the same columns name os the table that will be updated.
}
\examples{
\dontrun{
cl_new <- data.frame(
Identifier = c(999, 888),
Acronym = 'TEST_ACRONYM',
Folder = 'TESTFOLDER',
Name = 'TEST_NAME',
Branch = 'Fishery',
Instance = 'Fishery')

UpdateEBXCodeList(data     = cl_new,
                  sdmx_codelist_name  = 'EBXCodelist')
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
