% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_UpdateEBXCodeList.R
\name{UpdateEBXCodeList}
\alias{UpdateEBXCodeList}
\title{Update rows in EBX5 Code List}
\usage{
UpdateEBXCodeList(data, sdmx_name)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{sdmx_name}{code list name, in SDMX style. Please, see
the code lists available by running the function \code{\link{GetEBXCodeLists}} in the field "Acronym".
The actual codelist location in EBX5 (branch, instance, code-list-name) are resolved
using the metadata structure.}
}
\value{
boolean
}
\description{
This function aimed to update
data rows into a code list stored in EBX5 through R.
}
\details{
Note that the udpated rows must have the same columns name os the table that will be updated.
}
\examples{

\dontrun{
cl_new <- data.frame(
Identifier = c(999, 888),
Acronym = 'TEST_ACRONYM',
Folder = 'TESTFOLDER',
Name = 'TEST_NAME',
Branch = 'Fishery',
Instance = 'Fishery')

UpdateEBXCodeList(data     = cl_new,
                  cl_name  = 'EBXCodelist',
                  folder   = 'Metadata',
                  branch   = 'Fishery',
                  instance = 'Fishery')
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
