% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_UpdateEBXGroup.R
\name{UpdateEBXGroup}
\alias{UpdateEBXGroup}
\title{Update rows in EBX5 Group}
\usage{
UpdateEBXGroup(data, sdmx_name)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} that will be appended.}

\item{sdmx_name}{group name, in SDMX style. Please, see
the groups available by running the function \code{\link{GetEBXGroups}} in the field "Acronym".
The actual group location in EBX5 (branch, instance, code-list-name) are resolved
using the metadata structure.}
}
\value{
boolean
}
\description{
This function aimed to update
data rows into a group stored in EBX5 through R.
}
\details{
Note that the udpated rows must have the same columns name os the table that will be updated.
}
\examples{

\dontrun{
UpdateEBXGroup(data = gr_update,
gr_name = 'EBXGroup',
folder = 'Metadata')
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luis G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
